Chore: move folder, correct component name
- correct wrong component name
- move router to utils
