+++ src/components/Article/index.js
@@ -1,4 +1,4 @@
-import { Router } from "../../Router";
+import { Router } from "../../utils/router";
 import { Component } from "../Component";
 import "./index.css";
